Change file and folder structure of entire application in order to utilize the app router and then remove NavToPage state and replace with Link component for navigation.
