+++ src/app/about/AboutSection/AboutSection.css
@@ -1,0 +1,14 @@
+.image-container {
+  width: 10%;
+
+  > div {
+    position: unset !important;
+  }
+
+  .image {
+    object-fit: contain;
+    width: 100% !important;
+    position: relative !important;
+    height: unset !important;
+  }
+}
