--- src/app/components/Navbar/Navbar.tsx
@@ -1,39 +1,0 @@
-"use client";
-import DeskTopMenu from "./DesktopMenu/DeskTopMenu";
-import HamburgerMenu from "./HamburgerMenu/HamburgerMenu";
-import MobileMenu from "./MobileMenu/MobileMenu";
-import "./Navbar.css";
-import { useState } from "react";
-
-interface NavbarProps {
-  setNavToPage: (clicked: string) => void;
-}
-
-const Navbar: React.FC<NavbarProps> = ({ setNavToPage }) => {
-  const [showMobileMenu, setShowMobileMenu] = useState(false);
-
-  const handleClick = () => {
-    setNavToPage("home");
-  };
-
-  return (
-    <nav className="h-1/3">
-      <div className="grid grid-cols-2 gap-x-8 w-screen">
-        <button
-          className="logo flex justify-center items-end text-xl sm:text-2xl md:text-3xl lg:text-4xl xl:text-5xl"
-          onClick={handleClick}
-        >
-          c. lippman
-        </button>
-        <DeskTopMenu setNavToPage={setNavToPage} />
-        <HamburgerMenu
-          showMobileMenu={showMobileMenu}
-          setShowMobileMenu={setShowMobileMenu}
-        />
-      </div>
-      {showMobileMenu && <MobileMenu setNavToPage={setNavToPage} />}
-    </nav>
-  );
-};
-
-export default Navbar;
