--- src/app/components/WorkSection/WorkSection.css
@@ -1,14 +1,0 @@
-.image-container {
-  width: 100%;
-
-  > div {
-    position: unset !important;
-  }
-
-  .image {
-    object-fit: contain;
-    width: 100% !important;
-    position: relative !important;
-    height: unset !important;
-  }
-}
