+++ src/app/contact/ContactSection/ContactSection.tsx
@@ -1,0 +1,7 @@
+export default function ContactSection() {
+  return (
+    <main>
+      <h1>Contact Me</h1>
+    </main>
+  );
+}
